Rename auto_recreate_subscriptions to just recreate_subscriptions
This is of course again a nitpick (also the last one), but I felt
`auto_recreate_subscriptions` was a bit out of sync with the other
options used (as they are all also doing something automatically), like
`should_reconnect` and `recreate_monitored_items_chunk`.

In addition I (tried to) reorder the used options a bit so they are both
grouped and more or less in sync over places where a similar (or the
same) list of options are used in order to try and make it easier to
find things in the lists.

While I like these micro improvements, I get that different people may
have different preferences...