Relax loading XML values using FromValue
All the ones we currently use it for implement Default, so we can just
use unwrap_or_defualt. This handles the case of empty strings, and is
probably fine for other types. It's a little weird for numbers, but 0 is
a pretty reasonable way to interpret an empty <Int></Int> tag.